move the `WindowSpecification` definition.
