+++ pywinauto/application.py
@@ -30,7 +30,7 @@
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 import sys
 
-from .base_application import WindowSpecification  # noqa: W0611
+from .window_specification import WindowSpecification  # noqa: W0611
 
 if sys.platform == 'win32':
     from .windows.application import Application  # noqa: W0611
